Adding stuff

+++ api-gateway/src/Application/Metrics/CallData.cs
@@ -6,7 +6,7 @@ namespace Application.Metrics
 {
     public class CallData : MetricsBase
     {
-        public decimal Duration { get; set; }
+        public double Duration { get; set; }
         public string Status { get; set; }
         public string CallType { get; set; }
         public string Method { get; set; }
